cloud_functions/api-emails-1/requirements.txt
# Function dependencies, for example:
# package>=version

google-cloud-firestore